Moves factory enums to corresponding command packages
