+++ src/main/java/org/jboss/forge/arquillian/command/algeron/publisher/AlgeronPublishers.java
@@ -1,11 +1,8 @@
-package org.jboss.forge.arquillian.container.model;
+package org.jboss.forge.arquillian.command.algeron.publisher;
 
 import org.jboss.forge.addon.ui.command.UICommand;
-import org.jboss.forge.arquillian.command.algeron.publisher.AddAlgeronFolderPublisher;
-import org.jboss.forge.arquillian.command.algeron.publisher.AddAlgeronGitPublisher;
-import org.jboss.forge.arquillian.command.algeron.publisher.AddAlgeronUrlPublisher;
 
-public enum AlgeronPublishers {
+enum AlgeronPublishers {
 
     FOLDER(AddAlgeronFolderPublisher.class), URL(AddAlgeronUrlPublisher.class), GIT(AddAlgeronGitPublisher.class);
