+++ src/main/java/org/jboss/forge/arquillian/command/algeron/retriever/AlgeronRetriever.java
@@ -1,13 +1,8 @@
-package org.jboss.forge.arquillian.container.model;
+package org.jboss.forge.arquillian.command.algeron.retriever;
 
 import org.jboss.forge.addon.ui.command.UICommand;
-import org.jboss.forge.arquillian.command.algeron.retriever.AddAlgeronFolderRetriever;
-import org.jboss.forge.arquillian.command.algeron.retriever.AddAlgeronGitRetriever;
-import org.jboss.forge.arquillian.command.algeron.retriever.AddAlgeronMavenRetriever;
-import org.jboss.forge.arquillian.command.algeron.retriever.AddAlgeronPactBrokerRetriever;
-import org.jboss.forge.arquillian.command.algeron.retriever.AddAlgeronUrlRetriever;
 
-public enum AlgeronRetriever {
+enum AlgeronRetriever {
 
     FOLDER(AddAlgeronFolderRetriever.class), URL(AddAlgeronUrlRetriever.class), GIT(AddAlgeronGitRetriever.class),
     MAVEN(AddAlgeronMavenRetriever.class), PACT_BROKER(AddAlgeronPactBrokerRetriever.class);
